Building process has been mavenized. * Removed spigot.jar as library file. * Latest dependencies (for mc 1.8). * Package name changed. * Version hop to 1.1.
