--- src/plugin.yml
@@ -1,7 +1,0 @@
-name: MojangAPIProxy
-main: mojangapiproxy.MojangAPIProxy
-version: 1.0
-author: _Shevchik_
-commands:
-  mapiproxy:
-    description: MojangAPIProxyCommands
